backend/requirements.txt
google_cloud_aiplatform-1.40.dev20240123+llm.extension-py2.py3-none-any.whl[preview]
google-cloud-api-keys
google-cloud-resource-manager
google-cloud-secret-manager
google-cloud-storage
google-search-results==2.4.2
flask==3.0.0
flask-cors==4.0.0
langchain==0.0.354
pydantic
